chore(insights): split trends charts into per-chart folders
Split the flat charts/trends/ folder into charts/trends-bar-chart/ and
charts/trends-line-chart/ so each chart variant lives next to its own
transforms, stories, tests, and chart-specific click handlers.

shared/ is unchanged. Internal imports between moved files use ./ for siblings
and ../shared/... for the shared helpers — both new dirs sit at the same
depth, so no in-file import changes were needed beyond the two external
callers (Trends.tsx and funnelChartTransforms.ts).

Generated-By: PostHog Code
Task-Id: 779fb86d-40c5-4855-bb6a-d2f16013824d
