+++ products/product_analytics/frontend/charts/trends-bar-chart/TrendsBarChart.test.tsx
