+++ products/product_analytics/frontend/charts/trends-bar-chart/handleTrendsBarAggregatedChartClick.ts
