+++ products/product_analytics/frontend/charts/trends-bar-chart/trendsBarChartTransforms.test.ts
